requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.1
pydantic==2.12.5
numpy==2.3.5
scikit-learn==1.8.0
scipy==1.16.3
pandas==2.3.3